zerotrail-bug-api/api/requirements.txt
fastapi
pydantic
mangum
google-generativeai
uvicorn
requests
